Typo

+++ ulib/ml/FStar_Int64.ml
@@ -34,7 +34,7 @@ let int_to_int64 (x:Prims.int) : int64 = Int64.of_string (Prims.to_string x)
 
 let shift_right (a:int64) (b:uint32) : int64 = Int64.shift_right_logical a b
 let shift_left  (a:int64) (b:uint32) : int64 = Int64.shift_left a b
-let shift_arithmetic_right (a:int8) (b:uint32) : int64 = Int64.shift_right a b
+let shift_arithmetic_right (a:int64) (b:uint32) : int64 = Int64.shift_right a b
 
 (* Comparison operators *)
 let eq (a:int64) (b:int64) : bool = a = b
